+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - additional-flags input
 - [ISSUE 103] be able to access definition file from private repositories
 - [ISSUE 216] remove -project flag from build branch command
+- [ISSUE 217] add skip project execution (spe) feature
 
 ## Bugs